backend/requirements.txt
# FastAPI & Server
fastapi==0.115.6
uvicorn[standard]==0.34.0
websockets==14.1
python-multipart==0.0.18

# Database
sqlmodel==0.0.22
asyncpg==0.30.0
sqlalchemy[asyncio]==2.0.36
alembic==1.14.1

# Redis & Celery
redis==5.2.1
celery[redis]==5.4.0

# AI Models
anthropic==0.42.0
google-genai==1.7.0

# Git
gitpython==3.1.43
PyGithub==2.5.0

# Docker
docker==7.1.0

# Browser Automation
playwright==1.49.1

# Web Requests
httpx==0.28.1
aiohttp==3.11.11

# Utilities
python-dotenv==1.0.1
pydantic==2.10.3
pydantic-settings==2.7.0

# LangGraph
langgraph==0.2.60
langgraph-checkpoint==2.0.10
langgraph-checkpoint-postgres==2.0.11
langchain-core==0.3.28
